+++ .gitignore
@@ -5,6 +5,7 @@ out/
 .idea/
 UML.puml
 untitled.iml
+туц-туц.puml
 
 ### Eclipse ###
 .apt_generated
