--- src/exceptions/DoorStateException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DoorStateException extends Exception{
-    public DoorStateException(String message){
-        super(message);
-    }
-}
