+++ src/interfaces/HoldInterface.java
@@ -4,6 +4,5 @@ import people.Person;
 import things.Thing;
 
 public interface HoldInterface {
-    void holdSmth(Thing thing);
     void holdSmn(Person person);
 }
