--- src/people/Elly.java
@@ -1,10 +1,0 @@
-package people;
-
-import enumerations.*;
-
-public class Elly extends Person{
-    public Elly(Place place){
-            super(place);
-    }
-    public void showSmthTo(Person person, EllyShow thing){}
-}
