+++ src/people/Person.java
@@ -2,18 +2,18 @@ package people;
 
 import enumerations.*;
 import interfaces.*;
+import things.Coffin;
+import things.Scene;
 import things.Thing;
 
 public class Person implements PutInterface, HoldInterface, MovementInterface {
-    public Place place;
+    private Place place;
     private Emotion EmotionalState;
+    private int temperature;
 
-    protected Person(Place place) {
+    public Person(Place place, int temperature) {
         this.place = place;
-        Heart heart = new Heart(120);
-    }
-
-    public Person() {
+        this.temperature = temperature;
         Heart heart = new Heart(120);
     }
 
@@ -47,24 +47,11 @@ public class Person implements PutInterface, HoldInterface, MovementInterface {
         person.place = place;
     }
 
-    public void holdSmth(Thing thing) {
-        if (thing.weight > 3) {
-            class Friction {
-                public void causeDiscomfort() {
-                    thing.weight *= 2;
-                }
-            }
-            Friction fr = new Friction();
-            fr.causeDiscomfort();
-        }
-    }
+
 
     public void holdSmn(Person person) {
     }
 
-    public void goTo(Place place) {
-        this.place = place;
-    }
 
     public void jump() {
     }
@@ -89,6 +76,38 @@ public class Person implements PutInterface, HoldInterface, MovementInterface {
 
     public Emotion getEmotionalState() {
         return this.EmotionalState;
+    }
+    public void sleep(){}
+    public void stir(){feel(Emotion.DISCOMFORT);
+    }
+    public void showSmthTo(Person person, EllyShow thing){}
+    public void remember(Scene scene){}
+    public void kissSmn(Person person){}
+    public void goTo(Place place){
+        this.place = place;
+        if (place == Place.LADDER){
+            this.temperature += 1;
+        }
+        else{
+            this.temperature = 36;
+        }
+    }
+    public void lookAround(){}
+    public void clutch(Person gadge){
+        gadge.stir();
+    }
+    public void putBlanketOver(Person gadge){}
+    public void lightCoffinTops(){
+        Coffin.CoffinTop.levelOfLighting = 30;
+    }
+    public Coffin createCoffin(Thing wood){
+        Thing deck = new Thing(1){
+            public void setWeight(Thing wood){
+                wood.weight /= 1000;
+            }
+        };
+        deck.setWeight(1000);
+        return new Coffin(Place.EXHIBITION, 100, 9);
     }
 
 }
