+++ src/things/Door.java
@@ -1,7 +1,6 @@
 package things;
 
 import enumerations.*;
-import exceptions.DoorStateException;
 import exceptions.OpenDoorException;
 
 public class Door extends Thing{
@@ -10,25 +9,8 @@ public class Door extends Thing{
         super(place, quantity, weight);
         this.state = state;
     }
-    public DoorState getState() throws DoorStateException{
-        if ((state != DoorState.OPEN) && (state != DoorState.CLOSED)){
-            throw new DoorStateException("Так нельзя");
-        }
-        else{
-            return this.state;
-        }
-    }
-    public boolean isValidState(){
-        try{
-            getState();
-            System.out.println("normal");
-            return true;
-        }
-        catch (DoorStateException doorStateException){
-            System.out.println("ne normal");
-            return false;
-        }
-    }
+
+
 
     public void openEx(){
         if (this.state == DoorState.OPEN) {
