+++ src/things/Entity.java
@@ -1,10 +1,10 @@
 package things;
 
 import enumerations.Emotion;
-import people.Louis;
+import people.Person;
 
 public class Entity {
-    public void triggerLouis(Louis louis){
+    public void triggerLouis(Person louis){
         louis.feel(Emotion.FEAR);
         louis.feel(Emotion.DESPAIR);
         louis.feel(Emotion.SURPRISE);
